requirements.txt
pandas
SQLAlchemy>=2.0
requests
python-dotenv
